Add files via upload
This data set documents the party leadership succession in 23 parliamentary democracies(as defined by Lijphart 1999). There are 25 columns and 4559 rows in the data, it includes the country, party information, name, sex, and term information about the leaders, and it also includes a status vector which use one to indicate the leader is still in office and 0 to indicate that they are out of office. 
Source:
Horiuchi, Y., Laing, M., & Hart, P. ’t. (2015). Hard acts to follow: Predecessor effects on party leader survival. Party Politics, 21(3), 357–366. https://doi.org/10.1177/1354068812472577